fix git issues with AbstractTilesGetMapOutputFormat
